all but CF generates
